+++ dotnet-project-file-analyzers.slnx
@@ -4,5 +4,5 @@
   <Project Path="specs/Specs/Specs.csproj" />
   <Project Path="specs/TestData/TestData.csproj" />
   <Project Path="src/DotNetProjectFile.Analyzers/DotNetProjectFile.Analyzers.csproj" />
-  <Project Path="src/DotNetProjectFile.Diagnostics/DotNetProjectFile.Diagnostics.csproj" />
+  <Project Path="src/DotNetProjectFile.RuleCatalog/DotNetProjectFile.RuleCatalog.csproj" />
 </Solution>
