+++ src/DotNetProjectFile.RuleCatalog/Data/DiagnosticCollection.json
