+++ src/DotNetProjectFile.RuleCatalog/DotNetProjectFile.RuleCatalog.csproj
