+++ src/DotNetProjectFile.RuleCatalog/Extensions/System.String.cs
