+++ src/DotNetProjectFile.RuleCatalog/Json/NuGetVersionJsonConverter.cs
@@ -2,7 +2,7 @@ using NuGet.Versioning;
 using System.Text.Json;
 using System.Text.Json.Serialization;
 
-namespace DotNetProjectFile.Diagnostics.Json;
+namespace DotNetProjectFile.RuleCatalog.Json;
 
 internal sealed class NuGetVersionJsonConverter : JsonConverter<NuGetVersion?>
 {
