+++ src/DotNetProjectFile.RuleCatalog/Properties/GlobalUsings.cs
